# MNT: disable the coercion cache for the nogil build